requirements.txt
# usage: `pip install -r requirements.txt`
pip>=7.1.2

ply==3.10
networkx==2.4
scipy

pyaml
matplotlib
# easy extras
omega==0.3.1
gr1py==0.2.0

# extras (uncomment as needed)
# cvxopt==1.1.7  # for faster polytopic operations
# matplotlib  # for plots

# tests
nose
mock

# dev
gitpython==2.1.0
